requirements.txt
asgiref==3.9.1
django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
pyotp==2.9.0
qrcode==7.4.2
sqlparse==0.5.3
typing-extensions==4.14.1
